Randomized all actions on specx dynamic tables, added ability for regression test to test all iframes on page
