+++ spec/clients/specx_api.rb
@@ -1,7 +1,7 @@
 require "httparty"
 class SpecxApi
   include HTTParty
-  @@basebox_ip = "192.168.0.20"
+  @@basebox_ip = "192.168.0.36"
   base_uri "#{@@basebox_ip}/api"
 
   attr_accessor :options
@@ -11,7 +11,7 @@ class SpecxApi
   end
 
   def initialize
-    basebox_token = "EfwNzGS1gthMzyZo8yNt"
+    basebox_token = "kM72_o348kGsz1pxfzSB"
     self.options = {
       headers: {
         "Authorization": "token='#{basebox_token}'",
